+++ example/lib/presentation/user/pages/user_page.dart
@@ -1,0 +1,8 @@
+import 'package:y_lints/annotations.dart';
+
+import '../../_widget_stubs.dart';
+
+@Page()
+class UserPage extends StatelessWidget {
+  const UserPage();
+}