+++ example/lib/presentation/user/view/user_builder.dart
@@ -1,0 +1,8 @@
+import 'package:y_lints/annotations.dart';
+
+import '../../_widget_stubs.dart';
+
+@FeatureBuilder()
+class UserBuilder extends StatelessWidget {
+  const UserBuilder();
+}